This PR enforces 64-bit handling of memory when collecting from the
system, fixing a test failure on armhf systems.

Though in a real environment this shouldn't be an issue, our test cases
have mock 32 GB of memory, resulting in an overflow when running on
32-bit systems. Instead of creating unique test cases for this
architecture, this PR should allow for this theoretical situation to be
handled.

+++ insights/internal/collector/sysinfo/hardware/hardware.go
@@ -46,7 +46,7 @@ type gpu struct {
 
 // memory contains information for the system's memory.
 type memory struct {
-	Total int `json:"size"`
+	Total int64 `json:"size"`
 }
 
 // disk contains information of a disk or partition.
